+++ src/Components/App.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { BrowserRouter, Switch, Route } from 'react-router-dom';
-import HomePage from './Home';
+import HomePage from './Views/Home';
 import '../App.css';
 
 function App() {
